Adjust the owner of the async function service & the DAPR_HOST env var
